Add API tests

+++ reports/src/controllers/Report.ts
@@ -3,8 +3,13 @@ import mongoose from "mongoose";
 import Report from "../models/Report";
 
 const createReport = (req: Request, res: Response, next: NextFunction) => {
-  console.log(req.body);
   const { name, gender, body, category } = req.body;
+
+  if (!name || !gender || !body) {
+    return res
+      .status(400)
+      .json({ error: true, message: "Invalid data provided" });
+  }
 
   const report = new Report({
     _id: new mongoose.Types.ObjectId(),
@@ -15,36 +20,46 @@ const createReport = (req: Request, res: Response, next: NextFunction) => {
 
   return report
     .save()
-    .then((report) => res.status(201).json({ report }))
+    .then((report) => res.status(201).json({ error: false, report }))
     .catch((error) => res.status(500).json({ error }));
 };
 
 const getReport = (req: Request, res: Response, next: NextFunction) => {
   const reportId = req.params.reportId;
+  if (!mongoose.Types.ObjectId.isValid(reportId)) {
+    return res
+      .status(400)
+      .json({ error: true, message: "Invalid ID provided" });
+  }
 
   return Report.findById(reportId)
     .then((report) =>
       report
-        ? res.status(200).json({ report })
-        : res.status(404).json({ message: "Report not found" })
+        ? res.status(200).json({ error: false, report })
+        : res.status(404).json({ error: true, message: "Report not found" })
     )
     .catch((error) => res.status(500).json({ error }));
 };
 
 const getAll = (req: Request, res: Response, next: NextFunction) => {
   return Report.find()
-    .then((reports) => res.status(201).json({ reports }))
+    .then((reports) => res.status(200).json({ error: false, reports }))
     .catch((error) => res.status(500).json({ error }));
 };
 
 const deleteReport = (req: Request, res: Response, next: NextFunction) => {
   const reportId = req.params.reportId;
+  if (!mongoose.Types.ObjectId.isValid(reportId)) {
+    return res
+      .status(400)
+      .json({ error: true, message: "Invalid ID provided" });
+  }
 
   return Report.findByIdAndDelete(reportId)
     .then((report) =>
       report
-        ? res.status(201).json({ message: "Report Deleted" })
-        : res.status(404).json({ message: "Report not found" })
+        ? res.status(200).json({ error: false, message: "Report Deleted" })
+        : res.status(404).json({ error: true, message: "Report not found" })
     )
     .catch((error) => res.status(500).json({ error }));
 };
